+++ src/main/duke/java/Command.java
@@ -1,4 +1,4 @@
-package duke.java;
+package main.duke.java;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks,Ui ui, Storage storage);
