+++ src/main/duke/java/Deadline.java
@@ -1,4 +1,4 @@
-package duke.java;
+package main.duke.java;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
