+++ src/main/duke/java/Duke.java
@@ -1,4 +1,4 @@
-package duke.java;
+package main.duke.java;
 
 import java.io.IOException;
