+++ src/main/duke/java/ExitCommand.java
@@ -1,4 +1,4 @@
-package duke.java;
+package main.duke.java;
 
 public class ExitCommand extends Command{
