+++ src/main/duke/java/TaskList.java
@@ -1,6 +1,5 @@
-package duke.java;
+package main.duke.java;
 
-import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 
 public class TaskList {
